add ranges instruction to parser :eyes:

+++ src/global/protocol_structures/instructions.rs
@@ -691,6 +691,6 @@ pub struct TypeMetadata {
 #[derive(BinRead, BinWrite, Clone, Debug, PartialEq)]
 #[brw(little)]
 pub struct RangeData {
-    pub start: Integer,
-    pub end: Integer,
+    pub start: u8,
+    pub end: u8,
 }
